A stable version using float

+++ include/HX1.hpp
@@ -4,6 +4,6 @@
 #include "parameters.hpp"
 
 // Function to solve the heat exchanger 1 problem
-void HX1(double y_hx1[length_hx], double Ts_HX1_L, double Tss_HX1_0, int step, Parameters &params);
+void HX1(float y_hx1[length_hx], float Ts_HX1_L, float Tss_HX1_0, int step, Parameters &params);
 
 #endif // HX1_HPP
