+++ include/ode_solver_neutronics.hpp
@@ -3,9 +3,9 @@
 
 #include "parameters.hpp"
 
-typedef void (*OdeFuncPointer)(double, const double[length_neutr], double[length_neutr], Parameters &params, const double Keff[N]);
+typedef void (*OdeFuncPointer)(float, const float[length_neutr], float[length_neutr], Parameters &params, const float Keff[N]);
 
 // Update the signature to use VectorXd instead of state_type
-void ode_solver_neutr(double y[length_neutr], OdeFuncPointer ode_func, int step, Parameters &params, const double Keff[N]);
+void ode_solver_neutr(float y[length_neutr], OdeFuncPointer ode_func, int step, Parameters &params, const float Keff[N]);
 
 #endif // ODE_SOLVER_HPP
