+++ include/ode_solver_th.hpp
@@ -3,9 +3,9 @@
 
 #include "parameters.hpp"
 
-typedef void (*OdeFuncPointer_th)(double, const double[length_neutr], double[length_neutr], Parameters &params, const double q_prime[N]);
+typedef void (*OdeFuncPointer_th)(float, const float[length_neutr], float[length_neutr], Parameters &params, const float q_prime[N]);
 
 // Update the signature to use VectorXd instead of state_type
-void ode_solver_th(double y[length_th], OdeFuncPointer_th ode_func, int step, Parameters &params, const double q_prime[N]);
+void ode_solver_th(float y[length_th], OdeFuncPointer_th ode_func, int step, Parameters &params, const float q_prime[N]);
 
 #endif // ODE_SOLVER_HPP
