+++ src/HX1.cpp
@@ -2,13 +2,13 @@
 #include "parameters.hpp"
 #include "ode_solver_HX.hpp"
 
-void pde_to_ode_hx1(double t, const double y[length_hx], double dydt[length_hx], Parameters &params) {
+void pde_to_ode_hx1(float t, const float y[length_hx], float dydt[length_hx], Parameters &params) {
     // Split the input state vector y into u and v
-    const double* u = &y[0];    // u is the first half of y
-    const double* v = &y[Nx];   // v is the second half of y
+    const float* u = &y[0];    // u is the first half of y
+    const float* v = &y[Nx];   // v is the second half of y
 
-    double du_dt[Nx] = {0};    // Array to store the derivative of u
-    double dv_dt[Nx] = {0};    // Array to store the derivative of v
+    float du_dt[Nx] = {0};    // Array to store the derivative of u
+    float dv_dt[Nx] = {0};    // Array to store the derivative of v
 
     // Compute du_dt and dv_dt using the provided constants
     for (int i = 0; i < Nx; ++i) {
@@ -33,9 +33,9 @@ void pde_to_ode_hx1(double t, const double y[length_hx], double dydt[length_hx],
     }
 }
 
-void HX1(double y_hx1[length_hx], double Ts_HX1_L, double Tss_HX1_0, int step, Parameters &params) {
+void HX1(float y_hx1[length_hx], float Ts_HX1_L, float Tss_HX1_0, int step, Parameters &params) {
     // Set boundary conditions
-    double u[Nx], v[Nx];
+    float u[Nx], v[Nx];
     if (step == 0) {
         for (int i = 0; i < Nx; ++i) {
             u[i] = params.u_init[i];
@@ -51,7 +51,7 @@ void HX1(double y_hx1[length_hx], double Ts_HX1_L, double Tss_HX1_0, int step, P
     v[0] = Tss_HX1_0;
 
     // Initial condition vector y0
-    double y0[length_hx];
+    float y0[length_hx];
     if (step == 0) {
         for (int i = 0; i < Nx; ++i) {
             y0[i] = params.u_init[i];
